+++ tapo-py/src/handlers.rs
@@ -1,3 +1,4 @@
+mod child_devices;
 mod color_light_handler;
 mod generic_device_handler;
 mod hub_handler;
@@ -5,6 +6,7 @@ mod light_handler;
 mod plug_energy_monitoring_handler;
 mod plug_handler;
 
+pub use child_devices::*;
 pub use color_light_handler::*;
 pub use generic_device_handler::*;
 pub use hub_handler::*;