+++ tapo-py/src/handlers/child_devices.rs
@@ -1,0 +1,3 @@
+mod t31x_handler;
+
+pub use t31x_handler::*;