+++ tapo/src/responses/child_device_list_hub_result/ke100_result.rs
@@ -13,7 +13,7 @@ pub enum TemperatureUnitKE100 {
     Celsius,
 }
 
-/// KE100 thermostatic radiator valve (TRV).
+/// Device info of Tapo KE100 thermostatic radiator valve (TRV).
 ///
 /// Specific properties: `temperature_unit`, `current_temperature`, `target_temperature`,
 /// `min_control_temperature`, `max_control_temperature`, `temperature_offset`,